requirements.txt
python-telegram-bot==13.0
python-dotenv==1.0.1
urllib3==1.26.15